+++ Sources/Plasma/PubUtilLib/plParticleSystem/plParticleSystem.h
@@ -188,8 +188,8 @@ public:
     
     plParticleSDLMod* GetSDLMod() {return fParticleSDLMod;}
     // Functions related to/required by plModifier
-    int GetNumTargets() const override { return fTarget ? 1 : 0; }
-    plSceneObject* GetTarget(int w) const override { hsAssert(w < 1, "Bad target"); return fTarget; }
+    size_t GetNumTargets() const override { return fTarget ? 1 : 0; }
+    plSceneObject* GetTarget(size_t w) const override { hsAssert(w < 1, "Bad target"); return fTarget; }
     void AddTarget(plSceneObject* so) override;
     void RemoveTarget(plSceneObject* so) override;
