+++ Sources/Plasma/PubUtilLib/plSurface/plGrassShaderMod.h
@@ -73,8 +73,8 @@ public:
     void ResetWaves();
     void RefreshWaves();
 
-    int GetNumTargets() const override { return fTarget ? 1 : 0; }
-    plSceneObject* GetTarget(int w) const override { return fTarget; }
+    size_t GetNumTargets() const override { return fTarget ? 1 : 0; }
+    plSceneObject* GetTarget(size_t w) const override { return fTarget; }
     void AddTarget(plSceneObject *object) override;
     void RemoveTarget(plSceneObject *object) override;
